Refactor Boilerplate Identity (#8630)
