+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Server/Boilerplate.Server.Api/Controllers/Identity/IdentityController.PhoneConfirmation.cs
@@ -1,0 +1,82 @@
+﻿//+:cnd:noEmit
+using Humanizer;
+using Boilerplate.Server.Api.Services;
+using Boilerplate.Shared.Dtos.Identity;
+using Boilerplate.Server.Api.Models.Identity;
+
+namespace Boilerplate.Server.Api.Controllers.Identity;
+
+public partial class IdentityController
+{
+    [AutoInject] private SmsService smsService = default!;
+
+    [HttpPost]
+    public async Task SendConfirmPhoneToken(SendPhoneTokenRequestDto request, CancellationToken cancellationToken)
+    {
+        var user = await userManager.FindByPhoneNumber(request.PhoneNumber!)
+            ?? throw new BadRequestException(Localizer[nameof(AppStrings.UserNotFound)]);
+
+        if (await userManager.IsPhoneNumberConfirmedAsync(user))
+            throw new BadRequestException(Localizer[nameof(AppStrings.PhoneNumberAlreadyConfirmed)]);
+
+        await SendConfirmPhoneToken(user, cancellationToken);
+    }
+
+    [HttpPost, Produces<TokenResponseDto>()]
+    public async Task ConfirmPhone(ConfirmPhoneRequestDto request, CancellationToken cancellationToken)
+    {
+        var user = await userManager.FindByPhoneNumber(request.PhoneNumber!)
+            ?? throw new BadRequestException(Localizer[nameof(AppStrings.UserNotFound)]);
+
+        var expired = (DateTimeOffset.Now - user.PhoneNumberTokenRequestedOn) > AppSettings.Identity.PhoneNumberTokenLifetime;
+
+        if (expired)
+            throw new BadRequestException(nameof(AppStrings.ExpiredToken));
+
+        if (await userManager.IsLockedOutAsync(user))
+            throw new BadRequestException(Localizer[nameof(AppStrings.UserLockedOut), (DateTimeOffset.UtcNow - user.LockoutEnd!).Value.Humanize(culture: CultureInfo.CurrentUICulture)]);
+
+        var tokenIsValid = await userManager.VerifyUserTokenAsync(user, TokenOptions.DefaultPhoneProvider, FormattableString.Invariant($"VerifyPhoneNumber:{request.PhoneNumber},{user.PhoneNumberTokenRequestedOn?.ToUniversalTime()}"), request.Token!);
+
+        if (tokenIsValid is false)
+        {
+            await userManager.AccessFailedAsync(user);
+            throw new BadRequestException(nameof(AppStrings.InvalidToken));
+        }
+        await ((IUserPhoneNumberStore<User>)userStore).SetPhoneNumberConfirmedAsync(user, true, cancellationToken);
+        var result = await userManager.UpdateAsync(user);
+        if (result.Succeeded is false)
+            throw new ResourceValidationException(result.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
+
+        await ((IUserLockoutStore<User>)userStore).ResetAccessFailedCountAsync(user, cancellationToken);
+        user.PhoneNumberTokenRequestedOn = null; // invalidates phone token
+        var updateResult = await userManager.UpdateAsync(user);
+        if (updateResult.Succeeded is false)
+            throw new ResourceValidationException(updateResult.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
+
+        var token = await userManager.GenerateUserTokenAsync(user, TokenOptions.DefaultPhoneProvider, FormattableString.Invariant($"Otp,{user.OtpRequestedOn?.ToUniversalTime()}"));
+
+        await SignIn(new() { PhoneNumber = request.PhoneNumber, Otp = token }, cancellationToken);
+    }
+
+
+    private async Task SendConfirmPhoneToken(User user, CancellationToken cancellationToken)
+    {
+        var resendDelay = (DateTimeOffset.Now - user.PhoneNumberTokenRequestedOn) - AppSettings.Identity.PhoneNumberTokenLifetime;
+
+        if (resendDelay < TimeSpan.Zero)
+            throw new TooManyRequestsExceptions(Localizer[nameof(AppStrings.WaitForPhoneNumberTokenRequestResendDelay), resendDelay.Value.Humanize(culture: CultureInfo.CurrentUICulture)]);
+
+        user.PhoneNumberTokenRequestedOn = DateTimeOffset.Now;
+        var result = await userManager.UpdateAsync(user);
+
+        if (result.Succeeded is false)
+            throw new ResourceValidationException(result.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
+
+        var phoneNumber = user.PhoneNumber!;
+        var token = await userManager.GenerateUserTokenAsync(user, TokenOptions.DefaultPhoneProvider, FormattableString.Invariant($"VerifyPhoneNumber:{phoneNumber},{user.PhoneNumberTokenRequestedOn?.ToUniversalTime()}"));
+        var link = new Uri(HttpContext.Request.GetWebClientUrl(), $"{Urls.ConfirmPage}?phoneNumber={Uri.EscapeDataString(phoneNumber!)}&phoneToken={Uri.EscapeDataString(token)}&culture={CultureInfo.CurrentUICulture.Name}");
+
+        await smsService.SendSms(Localizer[nameof(AppStrings.ConfirmPhoneTokenSmsText), token], phoneNumber, cancellationToken);
+    }
+}
