+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Server/Boilerplate.Server.Api/Controllers/Identity/IdentityController.cs
@@ -1,7 +1,5 @@
 ﻿//+:cnd:noEmit
 using Humanizer;
-using Microsoft.AspNetCore.Components.Web;
-using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.BearerToken;
 //#if (signalr == true)
 using Microsoft.AspNetCore.SignalR;
@@ -18,9 +16,6 @@ namespace Boilerplate.Server.Api.Controllers.Identity;
 [Route("api/[controller]/[action]")]
 public partial class IdentityController : AppControllerBase, IIdentityController
 {
-    [AutoInject] private SmsService smsService = default!;
-    [AutoInject] private EmailService emailService = default!;
-    [AutoInject] private HtmlRenderer htmlRenderer = default!;
     [AutoInject] private IUserStore<User> userStore = default!;
     [AutoInject] private UserManager<User> userManager = default!;
     [AutoInject] private SignInManager<User> signInManager = default!;
@@ -85,108 +80,8 @@ public partial class IdentityController : AppControllerBase, IIdentityController
         }
     }
 
-    [HttpPost]
-    public async Task SendConfirmEmailToken(SendEmailTokenRequestDto request, CancellationToken cancellationToken)
-    {
-        var user = await userManager.FindByEmailAsync(request.Email!)
-            ?? throw new BadRequestException(Localizer[nameof(AppStrings.UserNotFound)]);
-
-        if (await userManager.IsEmailConfirmedAsync(user))
-            throw new BadRequestException(Localizer[nameof(AppStrings.EmailAlreadyConfirmed)]);
-
-        await SendConfirmEmailToken(user, cancellationToken);
-    }
-
-    [HttpPost]
-    public async Task<ActionResult<SignInResponseDto>> ConfirmEmail(ConfirmEmailRequestDto request, CancellationToken cancellationToken)
-    {
-        var user = await userManager.FindByEmailAsync(request.Email!)
-            ?? throw new BadRequestException(Localizer[nameof(AppStrings.UserNotFound)]);
-
-        var expired = (DateTimeOffset.Now - user.EmailTokenRequestedOn) > AppSettings.Identity.EmailTokenLifetime;
-
-        if (expired)
-            throw new BadRequestException();
-
-        if (await userManager.IsLockedOutAsync(user))
-            throw new BadRequestException(Localizer[nameof(AppStrings.UserLockedOut), (DateTimeOffset.UtcNow - user.LockoutEnd!).Value.Humanize(culture: CultureInfo.CurrentUICulture)]);
-
-        var tokenIsValid = await userManager.VerifyUserTokenAsync(user, TokenOptions.DefaultPhoneProvider, FormattableString.Invariant($"VerifyEmail:{request.Email},{user.EmailTokenRequestedOn?.ToUniversalTime()}"), request.Token!);
-
-        if (tokenIsValid is false)
-        {
-            await userManager.AccessFailedAsync(user);
-            throw new BadRequestException();
-        }
-
-        var userEmailStore = (IUserEmailStore<User>)userStore;
-        await userEmailStore.SetEmailConfirmedAsync(user, true, cancellationToken);
-        var result = await userManager.UpdateAsync(user);
-        if (result.Succeeded is false)
-            throw new ResourceValidationException(result.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
-
-        await ((IUserLockoutStore<User>)userStore).ResetAccessFailedCountAsync(user, cancellationToken);
-        user.EmailTokenRequestedOn = null; // invalidates email token
-        var updateResult = await userManager.UpdateAsync(user);
-        if (updateResult.Succeeded is false)
-            throw new ResourceValidationException(updateResult.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
-
-        var token = await userManager.GenerateUserTokenAsync(user, TokenOptions.DefaultPhoneProvider, FormattableString.Invariant($"Otp,{user.OtpRequestedOn?.ToUniversalTime()}"));
-
-        return await SignIn(new() { Email = request.Email, Otp = token }, cancellationToken);
-    }
-
-    [HttpPost]
-    public async Task SendConfirmPhoneToken(SendPhoneTokenRequestDto request, CancellationToken cancellationToken)
-    {
-        var user = await userManager.FindByPhoneNumber(request.PhoneNumber!)
-            ?? throw new BadRequestException(Localizer[nameof(AppStrings.UserNotFound)]);
-
-        if (await userManager.IsPhoneNumberConfirmedAsync(user))
-            throw new BadRequestException(Localizer[nameof(AppStrings.PhoneNumberAlreadyConfirmed)]);
-
-        await SendConfirmPhoneToken(user, cancellationToken);
-    }
-
-    [HttpPost]
-    public async Task<ActionResult<SignInResponseDto>> ConfirmPhone(ConfirmPhoneRequestDto request, CancellationToken cancellationToken)
-    {
-        var user = await userManager.FindByPhoneNumber(request.PhoneNumber!)
-            ?? throw new BadRequestException(Localizer[nameof(AppStrings.UserNotFound)]);
-
-        var expired = (DateTimeOffset.Now - user.PhoneNumberTokenRequestedOn) > AppSettings.Identity.PhoneNumberTokenLifetime;
-
-        if (expired)
-            throw new BadRequestException();
-
-        if (await userManager.IsLockedOutAsync(user))
-            throw new BadRequestException(Localizer[nameof(AppStrings.UserLockedOut), (DateTimeOffset.UtcNow - user.LockoutEnd!).Value.Humanize(culture: CultureInfo.CurrentUICulture)]);
-
-        var tokenIsValid = await userManager.VerifyUserTokenAsync(user, TokenOptions.DefaultPhoneProvider, FormattableString.Invariant($"VerifyPhoneNumber:{request.PhoneNumber},{user.PhoneNumberTokenRequestedOn?.ToUniversalTime()}"), request.Token!);
-
-        if (tokenIsValid is false)
-        {
-            await userManager.AccessFailedAsync(user);
-            throw new BadRequestException();
-        }
-        await ((IUserPhoneNumberStore<User>)userStore).SetPhoneNumberConfirmedAsync(user, true, cancellationToken);
-        var result = await userManager.UpdateAsync(user);
-        if (result.Succeeded is false)
-            throw new ResourceValidationException(result.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
-
-        await ((IUserLockoutStore<User>)userStore).ResetAccessFailedCountAsync(user, cancellationToken);
-        user.PhoneNumberTokenRequestedOn = null; // invalidates phone token
-        var updateResult = await userManager.UpdateAsync(user);
-        if (updateResult.Succeeded is false)
-            throw new ResourceValidationException(updateResult.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
-
-        var token = await userManager.GenerateUserTokenAsync(user, TokenOptions.DefaultPhoneProvider, FormattableString.Invariant($"Otp,{user.OtpRequestedOn?.ToUniversalTime()}"));
-
-        return await SignIn(new() { PhoneNumber = request.PhoneNumber, Otp = token }, cancellationToken);
-    }
-
-    [HttpPost]
-    public async Task<ActionResult<SignInResponseDto>> SignIn(SignInRequestDto request, CancellationToken cancellationToken)
+    [HttpPost, Produces<SignInResponseDto>()]
+    public async Task SignIn(SignInRequestDto request, CancellationToken cancellationToken)
     {
         signInManager.AuthenticationScheme = IdentityConstants.BearerScheme;
 
@@ -201,7 +96,7 @@ public partial class IdentityController : AppControllerBase, IIdentityController
             var expired = (DateTimeOffset.Now - user.OtpRequestedOn) > AppSettings.Identity.OtpTokenLifetime;
 
             if (expired)
-                throw new BadRequestException();
+                throw new BadRequestException(nameof(AppStrings.ExpiredToken));
         }
 
         var result = isOtpSignIn
@@ -230,7 +125,8 @@ public partial class IdentityController : AppControllerBase, IIdentityController
             }
             else
             {
-                return new SignInResponseDto { RequiresTwoFactor = true };
+                await Response.WriteAsJsonAsync(new SignInResponseDto { RequiresTwoFactor = true }, cancellationToken);
+                return;
             }
         }
 
@@ -254,8 +150,6 @@ public partial class IdentityController : AppControllerBase, IIdentityController
         //#if (signalr == true)
         await identityHubContext.Clients.User(user.Id.ToString()).SendAsync("NewUserSession", userSession, cancellationToken);
         //#endif
-
-        return Empty;
     }
 
     /// <summary>
@@ -336,50 +230,6 @@ public partial class IdentityController : AppControllerBase, IIdentityController
         return SignIn(newPrincipal, authenticationScheme: IdentityConstants.BearerScheme);
     }
 
-    [HttpPost]
-    public async Task SendResetPasswordToken(SendResetPasswordTokenRequestDto request, CancellationToken cancellationToken)
-    {
-        var user = await userManager.FindUserAsync(request)
-                    ?? throw new ResourceNotFoundException(Localizer[nameof(AppStrings.UserNotFound)]);
-
-        if (await userConfirmation.IsConfirmedAsync(userManager, user) is false)
-            throw new BadRequestException(Localizer[nameof(AppStrings.UserIsNotConfirmed)]);
-
-        var resendDelay = (DateTimeOffset.Now - user.ResetPasswordTokenRequestedOn) - AppSettings.Identity.ResetPasswordTokenLifetime;
-
-        if (resendDelay < TimeSpan.Zero)
-            throw new TooManyRequestsExceptions(Localizer[nameof(AppStrings.WaitForResetPasswordTokenRequestResendDelay), resendDelay.Value.Humanize(culture: CultureInfo.CurrentUICulture)]);
-
-        user.ResetPasswordTokenRequestedOn = DateTimeOffset.Now;
-
-        var result = await userManager.UpdateAsync(user);
-
-        if (result.Succeeded is false)
-            throw new ResourceValidationException(result.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
-
-        var token = await userManager.GenerateUserTokenAsync(user, TokenOptions.DefaultPhoneProvider, FormattableString.Invariant($"ResetPassword,{user.ResetPasswordTokenRequestedOn?.ToUniversalTime()}"));
-        var isEmail = string.IsNullOrEmpty(request.Email) is false;
-        var qs = $"{(isEmail ? "email" : "phoneNumber")}={Uri.EscapeDataString(isEmail ? request.Email! : request.PhoneNumber!)}";
-        var url = $"{Urls.ResetPasswordPage}?token={Uri.EscapeDataString(token)}&{qs}&culture={CultureInfo.CurrentUICulture.Name}";
-        var link = new Uri(HttpContext.Request.GetWebClientUrl(), url);
-
-        async Task SendEmail()
-        {
-            if (await userManager.IsEmailConfirmedAsync(user) is false) return;
-
-            await emailService.SendResetPasswordToken(user, token, link, cancellationToken);
-        }
-
-        async Task SendSms()
-        {
-            if (await userManager.IsPhoneNumberConfirmedAsync(user) is false) return;
-
-            await smsService.SendSms(Localizer[nameof(AppStrings.ResetPasswordTokenSmsText), token], user.PhoneNumber!, cancellationToken);
-        }
-
-        await Task.WhenAll([SendEmail(), SendSms()]);
-    }
-
     /// <summary>
     /// For either otp or magic link
     /// </summary>
@@ -419,39 +269,6 @@ public partial class IdentityController : AppControllerBase, IIdentityController
     }
 
     [HttpPost]
-    public async Task ResetPassword(ResetPasswordRequestDto request, CancellationToken cancellationToken)
-    {
-        var user = await userManager.FindUserAsync(request) ?? throw new ResourceNotFoundException(Localizer[nameof(AppStrings.UserNotFound)]);
-
-        var expired = (DateTimeOffset.Now - user.ResetPasswordTokenRequestedOn) > AppSettings.Identity.ResetPasswordTokenLifetime;
-
-        if (expired)
-            throw new BadRequestException();
-
-        if (await userManager.IsLockedOutAsync(user))
-            throw new BadRequestException(Localizer[nameof(AppStrings.UserLockedOut), (DateTimeOffset.UtcNow - user.LockoutEnd!).Value.Humanize(culture: CultureInfo.CurrentUICulture)]);
-
-        bool tokenIsValid = await userManager.VerifyUserTokenAsync(user!, TokenOptions.DefaultPhoneProvider, FormattableString.Invariant($"ResetPassword,{user.ResetPasswordTokenRequestedOn?.ToUniversalTime()}"), request.Token!);
-
-        if (tokenIsValid is false)
-        {
-            await userManager.AccessFailedAsync(user);
-            throw new BadRequestException();
-        }
-
-        var result = await userManager.ResetPasswordAsync(user!, await userManager.GeneratePasswordResetTokenAsync(user!), request.Password!);
-
-        if (result.Succeeded is false)
-            throw new ResourceValidationException(result.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
-
-        await ((IUserLockoutStore<User>)userStore).ResetAccessFailedCountAsync(user, cancellationToken);
-        user.ResetPasswordTokenRequestedOn = null; // invalidates reset password token
-        var updateResult = await userManager.UpdateAsync(user);
-        if (updateResult.Succeeded is false)
-            throw new ResourceValidationException(updateResult.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
-    }
-
-    [HttpPost]
     public async Task SendTwoFactorToken(IdentityRequestDto request, CancellationToken cancellationToken)
     {
         var user = await userManager.FindUserAsync(request) ?? throw new ResourceNotFoundException(Localizer[nameof(AppStrings.UserNotFound)]);
@@ -488,104 +305,12 @@ public partial class IdentityController : AppControllerBase, IIdentityController
     }
 
     [HttpGet]
-    public async Task<string> GetSocialSignInUri(string provider, string? returnUrl = null, int? localHttpPort = null, CancellationToken cancellationToken = default)
-    {
-        var uri = Url.Action(nameof(SocialSignIn), new { provider, returnUrl, localHttpPort })!;
-        return new Uri(Request.GetBaseUrl(), uri).ToString();
-    }
-
-    [HttpGet]
-    public async Task<ActionResult> SocialSignIn(string provider, string? returnUrl = null, int? localHttpPort = null)
-    {
-        var redirectUrl = Url.Action(nameof(SocialSignInCallback), "Identity", new { returnUrl, localHttpPort });
-        var properties = signInManager.ConfigureExternalAuthenticationProperties(provider, redirectUrl);
-        return new ChallengeResult(provider, properties);
-    }
-
-    [HttpGet]
-    public async Task<ActionResult> SocialSignInCallback(string? returnUrl = null, int? localHttpPort = null, CancellationToken cancellationToken = default)
-    {
-        string? url;
-
-        var info = await signInManager.GetExternalLoginInfoAsync() ?? throw new BadRequestException();
-
-        try
-        {
-            var email = info.Principal.GetEmail();
-            var phoneNumber = info.Principal.Claims.FirstOrDefault(c => c.Type is ClaimTypes.HomePhone or ClaimTypes.MobilePhone or ClaimTypes.OtherPhone)?.Value;
-
-            var user = await userManager.FindByLoginAsync(info.LoginProvider, info.ProviderKey);
-
-            if (user is null && (string.IsNullOrEmpty(email) is false || string.IsNullOrEmpty(phoneNumber) is false))
-            {
-                user = await userManager.FindUserAsync(new() { Email = email, PhoneNumber = phoneNumber });
-            }
-
-            if (user is null)
-            {
-                // Instead of automatically creating a user here, you can navigate to the sign-up page and pass the email and phone number in the query string.
-
-                user = new() { LockoutEnabled = true };
-
-                await userStore.SetUserNameAsync(user, Guid.NewGuid().ToString(), cancellationToken);
-
-                if (string.IsNullOrEmpty(email) is false)
-                {
-                    await ((IUserEmailStore<User>)userStore).SetEmailAsync(user, email, cancellationToken);
-                }
-
-                if (string.IsNullOrEmpty(phoneNumber) is false)
-                {
-                    await ((IUserPhoneNumberStore<User>)userStore).SetPhoneNumberAsync(user, phoneNumber!, cancellationToken);
-                }
-
-                var result = await userManager.CreateAsync(user, password: Guid.NewGuid().ToString("N") /* Users can reset their password later. */);
-
-                if (result.Succeeded is false)
-                {
-                    throw new BadRequestException(string.Join(", ", result.Errors.Select(e => new LocalizedString(e.Code, e.Description))));
-                }
-
-                await userManager.AddLoginAsync(user, info);
-            }
-
-            if (string.IsNullOrEmpty(email) is false && email == user.Email && await userManager.IsEmailConfirmedAsync(user) is false)
-            {
-                await ((IUserEmailStore<User>)userStore).SetEmailConfirmedAsync(user, true, cancellationToken);
-                await userManager.UpdateAsync(user);
-            }
-
-            if (string.IsNullOrEmpty(phoneNumber) is false && phoneNumber == user.PhoneNumber && await userManager.IsPhoneNumberConfirmedAsync(user) is false)
-            {
-                await ((IUserPhoneNumberStore<User>)userStore).SetPhoneNumberConfirmedAsync(user, true, cancellationToken);
-                await userManager.UpdateAsync(user);
-            }
-
-            (_, url) = await GenerateOtpTokenData(user, returnUrl); // Sign in with a magic link, and 2FA will be prompted if already enabled.
-        }
-        catch (Exception exp)
-        {
-            LogSocialSignInCallbackFailed(logger, exp, info.LoginProvider, info.Principal.GetDisplayName());
-            url = $"{Urls.SignInPage}?error={Uri.EscapeDataString(exp is KnownException ? Localizer[exp.Message] : Localizer[nameof(AppStrings.UnknownException)])}";
-        }
-        finally
-        {
-            await Request.HttpContext.SignOutAsync(IdentityConstants.ExternalScheme); // We'll handle sign-in with the following redirects, so no external identity cookie is needed.
-        }
-
-        if (localHttpPort is not null) return Redirect(new Uri(new Uri($"http://localhost:{localHttpPort}"), url).ToString());
-        var webClientUrl = Configuration.GetValue<string?>("WebClientUrl");
-        if (string.IsNullOrEmpty(webClientUrl) is false) return Redirect(new Uri(new Uri(webClientUrl), url).ToString());
-        return LocalRedirect($"~{url}");
-    }
-
-    [HttpGet]
     public async Task<ActionResult> SocialSignedIn()
     {
         var html = await htmlRenderer.Dispatcher.InvokeAsync(async () =>
                     (await htmlRenderer.RenderComponentAsync<SocialSignedInPage>()).ToHtmlString());
 
-        return Content(html, "text/html", System.Text.Encoding.UTF8);
+        return Content(html, "text/html");
     }
 
     [LoggerMessage(Level = LogLevel.Error, Message = "Failed to perform {loginProvider} social sign in for {principal}")]
@@ -612,45 +337,5 @@ public partial class IdentityController : AppControllerBase, IIdentityController
         var url = $"{Urls.SignInPage}?otp={Uri.EscapeDataString(token)}&{qs}&culture={CultureInfo.CurrentUICulture.Name}";
 
         return (token, url);
-    }
-
-    private async Task SendConfirmEmailToken(User user, CancellationToken cancellationToken)
-    {
-        var resendDelay = (DateTimeOffset.Now - user.EmailTokenRequestedOn) - AppSettings.Identity.EmailTokenLifetime;
-
-        if (resendDelay < TimeSpan.Zero)
-            throw new TooManyRequestsExceptions(Localizer[nameof(AppStrings.WaitForEmailTokenRequestResendDelay), resendDelay.Value.Humanize(culture: CultureInfo.CurrentUICulture)]);
-
-        user.EmailTokenRequestedOn = DateTimeOffset.Now;
-        var result = await userManager.UpdateAsync(user);
-
-        if (result.Succeeded is false)
-            throw new ResourceValidationException(result.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
-
-        var email = user.Email!;
-        var token = await userManager.GenerateUserTokenAsync(user, TokenOptions.DefaultPhoneProvider, FormattableString.Invariant($"VerifyEmail:{email},{user.EmailTokenRequestedOn?.ToUniversalTime()}"));
-        var link = new Uri(HttpContext.Request.GetWebClientUrl(), $"{Urls.ConfirmPage}?email={Uri.EscapeDataString(email)}&emailToken={Uri.EscapeDataString(token)}&culture={CultureInfo.CurrentUICulture.Name}");
-
-        await emailService.SendEmailToken(user, email, token, link, cancellationToken);
-    }
-
-    private async Task SendConfirmPhoneToken(User user, CancellationToken cancellationToken)
-    {
-        var resendDelay = (DateTimeOffset.Now - user.PhoneNumberTokenRequestedOn) - AppSettings.Identity.PhoneNumberTokenLifetime;
-
-        if (resendDelay < TimeSpan.Zero)
-            throw new TooManyRequestsExceptions(Localizer[nameof(AppStrings.WaitForPhoneNumberTokenRequestResendDelay), resendDelay.Value.Humanize(culture: CultureInfo.CurrentUICulture)]);
-
-        user.PhoneNumberTokenRequestedOn = DateTimeOffset.Now;
-        var result = await userManager.UpdateAsync(user);
-
-        if (result.Succeeded is false)
-            throw new ResourceValidationException(result.Errors.Select(e => new LocalizedString(e.Code, e.Description)).ToArray());
-
-        var phoneNumber = user.PhoneNumber!;
-        var token = await userManager.GenerateUserTokenAsync(user, TokenOptions.DefaultPhoneProvider, FormattableString.Invariant($"VerifyPhoneNumber:{phoneNumber},{user.PhoneNumberTokenRequestedOn?.ToUniversalTime()}"));
-        var link = new Uri(HttpContext.Request.GetWebClientUrl(), $"{Urls.ConfirmPage}?phoneNumber={Uri.EscapeDataString(phoneNumber!)}&phoneToken={Uri.EscapeDataString(token)}&culture={CultureInfo.CurrentUICulture.Name}");
-
-        await smsService.SendSms(Localizer[nameof(AppStrings.ConfirmPhoneTokenSmsText), token], phoneNumber, cancellationToken);
     }
 }
