+++ src/Templates/Boilerplate/Bit.Boilerplate/src/Server/Boilerplate.Server.Api/Mappers/IdentityMapper.cs
@@ -12,7 +12,5 @@ public static partial class IdentityMapper
 {
     public static partial UserDto Map(this User source);
     public static partial void Patch(this EditUserDto source, User destination);
-
-    // https://mapperly.riok.app/docs/configuration/before-after-map/
     public static partial UserSessionDto Map(this UserSession source);
 }
